Add layout design
