+++ src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter } from 'react-router-dom';
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import Header from './Components/Header/Header';
+import Layout from './Components/Layout/Layout';
 import AppRoutes from './routes/routes';
 import loadRockets from './redux/rockets/utils';
 import { fetchMissions } from './redux/missions/missionsActions';
@@ -14,8 +14,9 @@ function App() {
   }, []);
   return (
     <BrowserRouter>
-      <Header />
-      <AppRoutes />
+      <Layout>
+        <AppRoutes />
+      </Layout>
     </BrowserRouter>
   );
 }
